pyproject.toml
[project]
name = "event-photo-finder"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "insightface",
    "opencv-python-headless",
    "numpy",
    "python-multipart",
    "google-api-python-client",
    "google-auth",
    "python-dotenv",
    "tqdm",
    "Pillow",
    "onnxruntime",
]
